Enabled test cases of multiple structures generated from 1 input file
Now we can test the case where multiple structures are generated by one input file. The ligands are added from a SMILE file with the information of different ligands. Specifically the test case added tests 10 bidentate ligands. The test program will check (1) whether the expected number of structures are generate (2) do the names of the structure match the reference (3) Check each of the structures against the references for all the properties checked for single structure case.